From public to friend class in inheritance
I have previously choosen to make some members public in classes in
Optimizers.hpp in order to use single inheritance easily (is a + has a)
instead of virtual multiple inheritance, e.g. for Adam.

Now, to recover the correct interface without affecting inheritance,
that members are again private or protected, and Adam, which has an
RMSprop object, is declared as friend class in RMSprop.

Minor changes have been made in the main function of Adam to allow
this edit.